BP-1200: Don't assume the format of patterns in WS events
The code was assuming that the event data of an updated
record/document always contained its Name. But sometimes the
pattern may not be a subscription to that specific document, but
a more generic one (like Name:"*"), and so in that case, the first
pattern may not matched that assumption.

This commit updates the code to read the document name from the
event's "key" structure instead.